The big SvelteKit rewrite.
This is a huge commit, coming from another work-in-progress
repository. It represents a fairly ground-up rewrite of a lot of things,
as README describes:

* bring everything up to SvelteKit 1.x
* replace Javascript with Typescript throughout
* replace unpleasant attempts at object-orientation with a more ECMAScript-like code
  style, focused on modules exporting functions and POJOs.
* replace confusing (an understatement) message-passing, and focus on a more
  SvelteKit-like way of doing things: put functions into the modules or components
  calling them, directly manipulate stores, and assume stores are a global store of
  truth.

Many things are much better as a result. Past me was very, very wrong
trying to make things object-oriented. I have embraced my strongly-typed
POJO future.